\name{ClearData}
\alias{ClearData}
\title{A function to reset all the data values to NULL, for a new query.}
\description{A function to reset all the data values to NULL, for a new query.
The ClearData() function allows a user to reset the query parameters,
(start.date, metrics, etc) back to NULL.}
\value{Resets all the query parameters to NULL.}
