\name{GetAnyXMLAttribute}
\alias{GetAnyXMLAttribute}
\title{Function to return the value attribute of the nodes of the parsed XML.}
\usage{GetAnyXMLAttribute(vNode, attr.name)}
\description{Function to return the value attribute of the nodes of the parsed XML.}
\value{The value contained with the XML node.}
\arguments{\item{vNode}{The XML node to be inspected.}
\item{attr.name}{The attribute to be returned.}}
