\name{GetProfileXML}
\alias{GetProfileXML}
\title{Returns a profile XML for an account...}
\description{Returns a profile XML for an account
For each authenticated user for a GA, they may have multiple accounts,
We retrieve a raw XML of all the accounts an authenticated user has access
to.}
\value{The raw XML string from the GA API.}
