\name{Init}
\alias{Init}
\title{A function setting initial values of a GA URI query.}
\usage{Init(start.date, end.date, dimensions, metrics, segment, sort, filters,
    max.results, start.index, table.id)}
\description{A function setting initial values of a GA URI query.}
\value{Sets the initial query parameters.}
\arguments{\item{start.date}{See QueryBuilder()}
\item{end.date}{See QueryBuilder()}
\item{dimensions}{See QueryBuilder()}
\item{metrics}{See QueryBuilder()}
\item{segment}{See QueryBuilder()}
\item{sort}{See QueryBuilder()}
\item{filters}{See QueryBuilder()}
\item{max.results}{See QueryBuilder()}
\item{start.index}{See QueryBuilder()}
\item{table.id}{See QueryBuilder()}}
