\name{Validate}
\alias{Validate}
\title{Returns whether the Query has all the required parameters set.}
\description{Returns whether the Query has all the required parameters set. These are
the start.date, end.date, metrics, and table.id parameters.}
\value{TRUE if the query has all the required parameters. Otherwise stops the
program execution.}
